Fix: bug in timestamp

+++ src/utils/structures.rs
@@ -4,7 +4,7 @@ use std::collections::HashMap;
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Claims {
     pub sub: String,
-    pub exp: usize
+    pub exp: u64
 }
 
 pub struct Position {
